requirements/local.txt
aiohttp==3.8.6
aiosignal==1.3.1
alabaster==0.7.13
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astroid==2.15.8
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
binaryornot==0.4.4
black==23.9.1
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.4.0
coverage==7.3.2
crispy-bootstrap5==0.7
cryptography==41.0.4
cssbeautifier==1.14.9
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
Django==4.2.6
django-allauth==0.57.0
django-appconf==1.0.5
django-compressor==4.4
django-cors-headers==4.2.0
django-coverage-plugin==3.1.0
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.3.1
django-redis==5.4.0
django-stubs==4.2.4
django-stubs-ext==4.2.2
djangorestframework==3.14.0
djangorestframework-stubs==3.14.3
djlint==1.34.0
docutils==0.20.1
drf-spectacular==0.26.5
EditorConfig==0.12.3
executing==2.0.0
factory-boy==3.3.0
Faker==19.6.2
filelock==3.12.4
flake8==6.1.0
flake8-isort==6.1.0
frozenlist==1.4.0
gTTS==2.4.0
hiredis==2.2.3
html-tag-names==0.1.2
html-void-elements==0.1.0
identify==2.5.30
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.16.1
isort==5.12.0
jedi==0.19.1
Jinja2==3.1.2
jsbeautifier==1.14.9
json5==0.9.14
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
livereload==2.6.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy==1.5.1
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
openai==0.28.1
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.4.0
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg-binary==3.1.12
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytest==7.4.2
pytest-django==4.5.2
pytest-sugar==0.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rjsmin==1.2.1
rpds-py==0.10.4
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
stack-data==0.6.3
termcolor==2.3.0
text-unidecode==1.3
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
types-python-dateutil==2.8.19.14
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.8
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.24.5
watchdog==3.0.0
wcwidth==0.2.8
Werkzeug==3.0.0
whitenoise==6.5.0
wrapt==1.15.0
yarl==1.9.2
